Move init.php to config.php
