+++ functions.php
@@ -10,12 +10,11 @@
  * @link https://github.com/roots/sage/pull/1042
  */
 $sage_includes = [
-  'lib/init.php',                  // Initial theme setup and constants
-  'lib/wrapper.php',               // Theme wrapper class
-  'lib/config.php',                // Configuration
-  'lib/assets.php',                // Scripts and stylesheets
-  'lib/titles.php',                // Page titles
-  'lib/extras.php',                // Custom functions
+  'lib/assets.php',  // Scripts and stylesheets
+  'lib/config.php',  // Configuration
+  'lib/extras.php',  // Custom functions
+  'lib/titles.php',  // Page titles
+  'lib/wrapper.php'  // Theme wrapper class
 ];
 
 foreach ($sage_includes as $file) {
